move contractor/proto to renter/proto
