+++ modules/renter/contractor/persist.go
@@ -2,7 +2,7 @@ package contractor
 
 import (
 	"github.com/NebulousLabs/Sia/modules"
-	"github.com/NebulousLabs/Sia/modules/renter/contractor/proto"
+	"github.com/NebulousLabs/Sia/modules/renter/proto"
 	"github.com/NebulousLabs/Sia/types"
 )
