+++ modules/renter/pool.go
@@ -3,7 +3,7 @@ package renter
 import (
 	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/modules/renter/contractor"
-	"github.com/NebulousLabs/Sia/modules/renter/contractor/proto"
+	"github.com/NebulousLabs/Sia/modules/renter/proto"
 )
 
 // A hostPool is a collection of active host connections, in the form of
