+++ modules/renter/proto/editor.go
